refactor: Dao에 대한 인터페이스 추가
