+++ src/main/java/woowacourse/shoppingcart/service/CartService.java
@@ -18,7 +18,6 @@ import woowacourse.shoppingcart.exception.NotInCustomerCartItemException;
 @Service
 @Transactional(rollbackFor = Exception.class)
 public class CartService {
-
     private final CartItemDao cartItemDao;
     private final ProductDao productDao;
 
@@ -28,7 +27,7 @@ public class CartService {
     }
 
     public List<CartItemResponse> findCartItemsByCustomerId(final Long customerId) {
-        List<CartItemEntity> cartItemEntities = cartItemDao.findCartItemsByCustomerId(customerId);
+        List<CartItemEntity> cartItemEntities = cartItemDao.findAllByCustomerId(customerId);
         List<CartItem> cartItems = cartItemEntities.stream()
                 .map(this::convertEntityToCartItem)
                 .collect(Collectors.toList());
@@ -39,16 +38,16 @@ public class CartService {
     }
 
     private CartItem convertEntityToCartItem(CartItemEntity cartItemEntity) {
-        Product product = productDao.findProductById(cartItemEntity.getProductId());
+        Product product = productDao.findById(cartItemEntity.getProductId());
         return new CartItem(cartItemEntity.getId(), product, cartItemEntity.getQuantity());
     }
 
     public Long addCart(final CartItemRequest cartItemRequest, final Long customerId) {
-        Product product = productDao.findProductById(cartItemRequest.getProductId());
+        Product product = productDao.findById(cartItemRequest.getProductId());
         CartItem cartItem = new CartItem(customerId, product, cartItemRequest.getQuantity());
 
         try {
-            return cartItemDao.addCartItem(customerId, cartItem);
+            return cartItemDao.save(customerId, cartItem);
         } catch (Exception e) {
             throw new InvalidProductException();
         }
@@ -56,29 +55,29 @@ public class CartService {
 
     public void deleteCart(final Long customerId, final Long cartItemId) {
         validateCustomerCart(cartItemId, customerId);
-        cartItemDao.deleteCartItem(cartItemId);
+        cartItemDao.delete(cartItemId);
     }
 
     public void updateCartItem(Long customerId, Long cartItemId, CartItemRequest cartItemRequest) {
         validateCustomerCart(cartItemId, customerId);
         validateProductId(cartItemId, cartItemRequest.getProductId());
 
-        Product newProduct = productDao.findProductById(cartItemRequest.getProductId());
+        Product newProduct = productDao.findById(cartItemRequest.getProductId());
         Integer newQuantity = cartItemRequest.getQuantity();
         CartItem newCartItem = new CartItem(cartItemId, newProduct, newQuantity);
 
-        cartItemDao.updateCartItem(cartItemId, newCartItem);
+        cartItemDao.update(cartItemId, newCartItem);
     }
 
     private void validateCustomerCart(final Long cartItemId, final Long customerId) {
-        CartItemEntity cartItemEntity = cartItemDao.findCartItemById(cartItemId);
+        CartItemEntity cartItemEntity = cartItemDao.findById(cartItemId);
         if (!cartItemEntity.getCustomerId().equals(customerId)) {
             throw new NotInCustomerCartItemException();
         }
     }
 
     private void validateProductId(final Long cartItemId, final Long productId) {
-        CartItemEntity cartItemEntity = cartItemDao.findCartItemById(cartItemId);
+        CartItemEntity cartItemEntity = cartItemDao.findById(cartItemId);
         if (!cartItemEntity.getProductId().equals(productId)) {
             throw new IllegalArgumentException("productId는 변경될 수 없습니다.");
         }
