+++ src/main/java/woowacourse/shoppingcart/service/ProductService.java
@@ -16,7 +16,7 @@ public class ProductService {
     }
 
     public List<Product> findProducts() {
-        return productDao.findProducts();
+        return productDao.findAll();
     }
 
     public Long addProduct(final Product product) {
@@ -24,7 +24,7 @@ public class ProductService {
     }
 
     public Product findProductById(final Long productId) {
-        return productDao.findProductById(productId);
+        return productDao.findById(productId);
     }
 
     public void deleteProductById(final Long productId) {
